requirements.txt
aiogram==3.4.1
APScheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.19.0